refactored names and function, and unit tests

+++ lib/src/modules/contact/domain/usecases/delete_user.dart
@@ -4,7 +4,7 @@ import '../errors/errors.dart';
 import '../repositories/user_repository.dart';
 
 abstract class IDeleteUser {
-  Future<Either<IUserException, void>> call({required int userId});
+  Future<Either<IUserException, int>> call({required final int userId});
 }
 
 class DeleteUser implements IDeleteUser {
@@ -13,7 +13,11 @@ class DeleteUser implements IDeleteUser {
   DeleteUser(this.repository);
 
   @override
-  Future<Either<IUserException, void>> call({required int userId}) async {
+  Future<Either<IUserException, int>> call({required final int userId}) async {
+    if (userId < 1) {
+      return left(const ArgumentsException('userId is invalid'));
+    }
+
     return await repository.deleteUser(userId: userId);
   }
 }
